requirements.txt
django==2.2.1
django-admin==1.3.2
mysqlclient
ipdb
djangorestframework
markdown
django-filter